# Perf/eth getlogs with compact encoding