--- a/src/Nethermind/Nethermind.Blockchain/Receipts/KeccaksIterator.cs
+++ b/src/Nethermind/Nethermind.Blockchain/Receipts/KeccaksIterator.cs
@@ -11,12 +11,15 @@ namespace Nethermind.Blockchain.Receipts
     {
         private readonly int _length;
         private Rlp.ValueDecoderContext _decoderContext;
+        private readonly Span<byte> _buffer;
         public long Index { get; private set; }
 
-        public KeccaksIterator(Span<byte> data)
+        public KeccaksIterator(Span<byte> data, Span<byte> buffer)
         {
+            if (buffer.Length != 32) throw new ArgumentException("Buffer must be 32 bytes long");
             _decoderContext = new Rlp.ValueDecoderContext(data);
             _length = _decoderContext.ReadSequenceLength();
+            _buffer = buffer;
             Index = -1;
         }
 
@@ -24,7 +27,7 @@ namespace Nethermind.Blockchain.Receipts
         {
             if (_decoderContext.Position < _length)
             {
-                _decoderContext.DecodeKeccakStructRef(out current);
+                _decoderContext.DecodeZeroPrefixedKeccakStructRef(out current, _buffer);
                 Index++;
                 return true;
             }