--- a/src/Nethermind/Nethermind.Blockchain/Receipts/LogEntriesIterator.cs
+++ b/src/Nethermind/Nethermind.Blockchain/Receipts/LogEntriesIterator.cs
@@ -12,14 +12,16 @@ namespace Nethermind.Blockchain.Receipts
         private readonly LogEntry[]? _logs;
         private readonly int _length;
         private Rlp.ValueDecoderContext _decoderContext;
+        private IReceiptRefDecoder _receiptRefDecoder;
         public long Index { get; private set; }
 
-        public LogEntriesIterator(Span<byte> data)
+        public LogEntriesIterator(Span<byte> data, IReceiptRefDecoder receiptRefDecoder)
         {
             _decoderContext = new Rlp.ValueDecoderContext(data);
             _length = _decoderContext.ReadSequenceLength();
             Index = -1;
             _logs = null;
+            _receiptRefDecoder = receiptRefDecoder;
         }
 
         public LogEntriesIterator(LogEntry[] logs)
@@ -36,7 +38,7 @@ namespace Nethermind.Blockchain.Receipts
             {
                 if (_decoderContext.Position < _length)
                 {
-                    LogEntryDecoder.DecodeStructRef(ref _decoderContext, RlpBehaviors.None, out current);
+                    _receiptRefDecoder.DecodeLogEntryStructRef(ref _decoderContext, RlpBehaviors.None, out current);
                     Index++;
                     return true;
                 }