--- a/src/Nethermind/Nethermind.Blockchain/Receipts/PersistentReceiptStorage.cs
+++ b/src/Nethermind/Nethermind.Blockchain/Receipts/PersistentReceiptStorage.cs
@@ -181,17 +181,22 @@ namespace Nethermind.Blockchain.Receipts
 
             var result = CanGetReceiptsByHash(blockNumber);
             var receiptsData = _blocksDb.GetSpan(blockHash);
-            IReceiptsRecovery.IRecoveryContext? recoveryContext = null;
+
+
+            Func<IReceiptsRecovery.IRecoveryContext?> recoveryContextFactory = () => null;
 
             if (_storageDecoder.IsCompactEncoding(receiptsData))
             {
-                Block block = _blockTree.FindBlock(blockHash);
-                recoveryContext = _receiptsRecovery.CreateRecoveryContext(block!);
+                recoveryContextFactory = () =>
+                {
+                    Block block = _blockTree.FindBlock(blockHash);
+                    return _receiptsRecovery.CreateRecoveryContext(block!);
+                };
             }
 
             IReceiptRefDecoder refDecoder = _storageDecoder.GetRefDecoder(receiptsData);
 
-            iterator = result ? new ReceiptsIterator(receiptsData, _blocksDb, recoveryContext, refDecoder) : new ReceiptsIterator();
+            iterator = result ? new ReceiptsIterator(receiptsData, _blocksDb, recoveryContextFactory, refDecoder) : new ReceiptsIterator();
             return result;
         }