--- a/src/Nethermind/Nethermind.Blockchain/Receipts/ReceiptsIterator.cs
+++ b/src/Nethermind/Nethermind.Blockchain/Receipts/ReceiptsIterator.cs
@@ -16,19 +16,25 @@ namespace Nethermind.Blockchain.Receipts
         private readonly IDbWithSpan _blocksDb;
         private readonly int _length;
         private Rlp.ValueDecoderContext _decoderContext;
+        private int _startingPosition;
 
         private readonly TxReceipt[]? _receipts;
-        private int _position;
-        private readonly IReceiptsRecovery.IRecoveryContext? _recoveryContext;
-        private IReceiptRefDecoder _receiptRefDecoder;
+        private int _receiptIndex;
 
-        public ReceiptsIterator(scoped in Span<byte> receiptsData, IDbWithSpan blocksDb, IReceiptsRecovery.IRecoveryContext? receiptsRecoveryContext, IReceiptRefDecoder receiptRefDecoder)
+        private readonly Func<IReceiptsRecovery.IRecoveryContext>? _recoveryContextFactory;
+        private IReceiptsRecovery.IRecoveryContext? _recoveryContext;
+        private IReceiptRefDecoder _receiptRefDecoder;
+        private bool _recoveryContextConfigured;
+
+        public ReceiptsIterator(scoped in Span<byte> receiptsData, IDbWithSpan blocksDb, Func<IReceiptsRecovery.IRecoveryContext?>? recoveryContextFactory, IReceiptRefDecoder receiptRefDecoder)
         {
             _decoderContext = receiptsData.AsRlpValueContext();
             _blocksDb = blocksDb;
             _receipts = null;
-            _position = 0;
-            _recoveryContext = receiptsRecoveryContext;
+            _receiptIndex = 0;
+            _recoveryContextFactory = recoveryContextFactory;
+            _recoveryContextConfigured = false;
+            _recoveryContext = null;
             _receiptRefDecoder = receiptRefDecoder;
 
             if (_decoderContext.Length > 0 && _decoderContext.PeekByte() == ReceiptArrayStorageDecoder.CompactEncoding)
@@ -36,6 +42,7 @@ namespace Nethermind.Blockchain.Receipts
                 _decoderContext.ReadByte();
             }
 
+            _startingPosition = _decoderContext.Position;
             _length = receiptsData.Length == 0 ? 0 : _decoderContext.ReadSequenceLength();
         }
 
@@ -49,7 +56,8 @@ namespace Nethermind.Blockchain.Receipts
             _length = receipts.Length;
             _blocksDb = null;
             _receipts = receipts;
-            _position = 0;
+            _receiptIndex = 0;
+            _recoveryContextConfigured = true;
         }
 
         public bool TryGetNext(out TxReceiptStructRef current)
@@ -60,14 +68,15 @@ namespace Nethermind.Blockchain.Receipts
                 {
                     _receiptRefDecoder.DecodeStructRef(ref _decoderContext, RlpBehaviors.Storage, out current);
                     _recoveryContext?.RecoverReceiptData(ref current);
+                    _receiptIndex++;
                     return true;
                 }
             }
             else
             {
-                if (_position < _length)
+                if (_receiptIndex < _length)
                 {
-                    current = new TxReceiptStructRef(_receipts[_position++]);
+                    current = new TxReceiptStructRef(_receipts[_receiptIndex++]);
                     return true;
                 }
             }
@@ -76,17 +85,24 @@ namespace Nethermind.Blockchain.Receipts
             return false;
         }
 
-        public void Reset()
+        public void RecoverIfNeeded(ref TxReceiptStructRef current)
         {
-            if (_receipts is not null)
+            if (_recoveryContextConfigured) return;
+
+            _recoveryContext = _recoveryContextFactory?.Invoke();
+            if (_recoveryContext != null)
             {
-                _position = 0;
+                // Need to replay the context.
+                _decoderContext.Position = _startingPosition;
+                if (_length != 0) _decoderContext.ReadSequenceLength();
+                for (int i = 0; i < _receiptIndex; i++)
+                {
+                    _receiptRefDecoder.DecodeStructRef(ref _decoderContext, RlpBehaviors.Storage, out current);
+                    _recoveryContext?.RecoverReceiptData(ref current);
+                }
             }
-            else
-            {
-                _decoderContext.Position = 0;
-                _decoderContext.ReadSequenceLength();
-            }
+
+            _recoveryContextConfigured = true;
         }
 
         public void Dispose()