--- a/src/Nethermind/Nethermind.Facade/Filters/Topics/AnyTopicsFilter.cs
+++ b/src/Nethermind/Nethermind.Facade/Filters/Topics/AnyTopicsFilter.cs
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2022 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
+using System;
 using Nethermind.Blockchain.Receipts;
 using Nethermind.Core;
 using Nethermind.Core.Crypto;
@@ -42,7 +43,8 @@ namespace Nethermind.Blockchain.Filters.Topics
                 return Accepts(entry.Topics);
             }
 
-            var iterator = new KeccaksIterator(entry.TopicsRlp);
+            Span<byte> buffer = stackalloc byte[32];
+            var iterator = new KeccaksIterator(entry.TopicsRlp, buffer);
             for (int i = 0; i < _expressions.Length; i++)
             {
                 if (iterator.TryGetNext(out var keccak))