--- a/src/Nethermind/Nethermind.Facade/Filters/Topics/SequenceTopicsFilter.cs
+++ b/src/Nethermind/Nethermind.Facade/Filters/Topics/SequenceTopicsFilter.cs
@@ -47,7 +47,8 @@ namespace Nethermind.Blockchain.Filters.Topics
                 return Accepts(entry.Topics);
             }
 
-            var iterator = new KeccaksIterator(entry.TopicsRlp);
+            Span<byte> buffer = stackalloc byte[32];
+            var iterator = new KeccaksIterator(entry.TopicsRlp, buffer);
             for (int i = 0; i < _expressions.Length; i++)
             {
                 if (iterator.TryGetNext(out var keccak))