--- a/src/Nethermind/Nethermind.Serialization.Rlp/CompactLogEntryDecoder.cs
+++ b/src/Nethermind/Nethermind.Serialization.Rlp/CompactLogEntryDecoder.cs
@@ -67,6 +67,43 @@ namespace Nethermind.Serialization.Rlp
             return new LogEntry(address, data, topics.ToArray());
         }
 
+        public void DecodeLogEntryStructRef(scoped ref Rlp.ValueDecoderContext decoderContext, RlpBehaviors behaviors, out LogEntryStructRef item)
+        {
+            if (decoderContext.IsNextItemNull())
+            {
+                decoderContext.ReadByte();
+                item = new LogEntryStructRef();
+                return;
+            }
+
+            decoderContext.ReadSequenceLength();
+            decoderContext.DecodeAddressStructRef(out var address);
+            var peekPrefixAndContentLength = decoderContext.PeekPrefixAndContentLength();
+            var sequenceLength = peekPrefixAndContentLength.PrefixLength + peekPrefixAndContentLength.ContentLength;
+            var topics = decoderContext.Data.Slice(decoderContext.Position, sequenceLength);
+            decoderContext.SkipItem();
+
+            int zeroPrefix = decoderContext.DecodeInt();
+            ReadOnlySpan<byte> rlpData = decoderContext.DecodeByteArraySpan();
+            byte[] data = new byte[zeroPrefix + rlpData.Length];
+            rlpData.CopyTo(data.AsSpan(zeroPrefix));
+
+            item = new LogEntryStructRef(address, data, topics);
+        }
+
+        public Keccak[] DecodeTopics(Rlp.ValueDecoderContext valueDecoderContext)
+        {
+            long sequenceLength = valueDecoderContext.ReadSequenceLength();
+            long untilPosition = valueDecoderContext.Position + sequenceLength;
+            using ArrayPoolList<Keccak> topics = new((int)(sequenceLength * 2 / Rlp.LengthOfKeccakRlp));
+            while (valueDecoderContext.Position < untilPosition)
+            {
+                topics.Add(valueDecoderContext.DecodeZeroPrefixKeccak());
+            }
+
+            return topics.ToArray();
+        }
+
         public void Encode(RlpStream rlpStream, LogEntry? item, RlpBehaviors rlpBehaviors = RlpBehaviors.None)
         {
             if (item is null)