--- a/src/Nethermind/Nethermind.Serialization.Rlp/CompactReceiptStorageDecoder.cs
+++ b/src/Nethermind/Nethermind.Serialization.Rlp/CompactReceiptStorageDecoder.cs
@@ -156,7 +156,7 @@ namespace Nethermind.Serialization.Rlp
         public void DecodeStructRef(scoped ref Rlp.ValueDecoderContext decoderContext, RlpBehaviors rlpBehaviors,
             out TxReceiptStructRef item)
         {
-            // Note: This method runs at 1.5 million times/sec on my machine
+            // Note: This method runs at 2.5 million times/sec on my machine
             item = new TxReceiptStructRef();
 
             if (decoderContext.IsNextItemNull())