… src/Nethermind/Nethermind.Serialization.Rlp/IReceiptRefDecoder.cs
… src/Nethermind/Nethermind.Serialization.Rlp/ReceiptArrayStorageDecoder.cs
… src/Nethermind/Nethermind.Serialization.Rlp/ReceiptStorageDecoder.cs
… src/Nethermind/Nethermind.Serialization.Rlp/Rlp.cs