- Optimize eth_getLogs on case where it have filter such as CL deposit scans.
- Skip bloom generation entirely on decoding refstruct. LogFinder skips bloom check if encoder does not support bloom decoding.
  - Surprisingly on CL workload this is faster than even with encoded flat bloom. 
- Delay recovery until the receipt match filter. 30% of receipt does not match, so this save block loading by that much.
- Improve max req per second (on same benchmark as before) from 2900 to 5700. Max io increase from 650MBps to 1200MBps. 95p latency down from 68ms to 41ms. 50p from 30ms to 13ms.

## Types of changes

#### What types of changes does your code introduce?

- [X] Optimization

## Testing

#### Requires testing

- [X] Yes
- [ ] No

#### If yes, did you write tests?

- [X] Yes
- [ ] No